test: extract tool registry white-box suite
